Always reuse DerivationStrategyFactory from the NBX Network instead of creating a new instance
If there is an NBX network that has its own derived factory, NBX would fail in various places to use that correct type. This fixes that but also makes sense beyond that. Also included here: typo fix in a log + helper method to create `ExplorerClient` from the NBX network directly